extract get libraries and re-render after create
