+++ web/src/app/library/library-workspace/library-workspace.component.html
@@ -1,5 +1,5 @@
 <div class="container">
   <app-library-create (create)="createLibrary($event)"></app-library-create>
   <hr />
-  <app-library-select></app-library-select>
+  <app-library-select [libraryData]="libraryData"></app-library-select>
 </div>
